DEBUG: params for sdc
