+++ deepprep/web/pages/1_🚀Preprocessing of T1w & BOLD.py
@@ -140,12 +140,16 @@ if selected_option != "T1w only":
         bold_sdc = st.checkbox("bold_sdc", value=True,
                                help="applies susceptibility distortion correction (SDC), default is True.")
         if bold_sdc:
-            deepprep_cmd += ' --bold_sdc'
+            deepprep_cmd += ' --bold_sdc True'
+        else:
+            deepprep_cmd += ' --bold_sdc False'
     with col5:
         bold_confounds = st.checkbox("bold_confounds", value=True,
                                      help="generates confounds derived from BOLD fMRI, such as head motion variables and global signals, default is True.")
         if bold_confounds:
-            deepprep_cmd += ' --bold_confounds'
+            deepprep_cmd += ' --bold_confounds True'
+        else:
+            deepprep_cmd += ' --bold_confounds False'
 
 participant_label = st.text_input("the subject IDs (optional)", placeholder="sub-001 sub-002 sub-003",
                                   help="Identify the subjects you'd like to process by their IDs, i.e. 'sub-001 sub-002'.")
